refactor: small restructuring to avoid madge syntax error on AppCore.vue
