+++ frontend/app/src/composables/use-core-scroll.ts
@@ -1,0 +1,14 @@
+export function useCoreScroll() {
+  const { y: scrollY } = useScroll(document.body);
+
+  const shouldShowScrollToTopButton = computed<boolean>(() => get(scrollY) > 200);
+
+  function scrollToTop() {
+    document.body.scrollTo(0, 0);
+  }
+
+  return {
+    scrollToTop,
+    shouldShowScrollToTopButton,
+  };
+}
